Chapter 1: change colors and sizes for SVG sprites
The size of the text map, 60x25, had narrow tall tiles, but the
SVG map has square tiles. That meant the new tiles were roughly half
the size of the old ones. It's harder to read the new icons at that
size, so I increased the size (by shrinking the map) and also changed
the color scheme to increase contrast.