Add types to much of the code
At first I planned to ignore any additional type errors but I decided
to clean up some of them, and ended up cleaning up most of them.

One of the trickier bits was the Location type, which is a union of a
location on the map {x,y}, being carried by someone, or being equipped
by someone. There are places that *I* know the location is on the map,
but the type system doesn't, so I can't pass a Location to a function
that wants a Point.

To solve that I parameterized the Entity so that it's either an Entity
that has a location, or an entity that has a location on the map. Then
when I .filter(…) the entities to return only the ones on the map, I
use .filter<EntityOnMap>(…) so that it returns an EntityOnMap[].
Accessing the location from within there will return a Point and not
an arbitary Location.

Unlike Rust, Typescript can't *guarantee* this will work, but it seems
reasonable. The way it works is that Typescript has a feature called
"type predicates"[1] that let it understand a type checking function
f(x). If f(x:T):boolean whether the type of x is U, then you can tell
Typescript about this by giving it the type f(x:T):x is U. Then in the
code if(f(x)) { … }, Typescript will give x type U inside the 'then'
clause and type T-U in the 'else' clause ("narrowing").

[1]: https://www.typescriptlang.org/docs/handbook/2/narrowing.html#using-type-predicates