Add CSS transitions to move sprites around smoothly
To implement this I needed to reuse the SVG DOM nodes, and I also
needed to segregate the sprites into layers so that they'd be drawn in
the right order (e.g. monster always on top of a potion, potion on top
of a corpse).

I already had 6 layers defined for sorting purposes so I reused those
for the drawing layers.