Move entities to a separate module
This required moving all the entity functions from being
top-level (which made it easier to keep data separate from functions)
into being methods (which is part of the reason I removed save/load).
Importing a bunch of top-level functions seems to be much more of a
pain with ES6 modules.