Feat: inventory items can stack, and direct use of them from ui
In the previous change, inventory excluded equippable items. Now
inventory items are only consumables. In this change I made them
stack. The inventory data structure, instead of being an array
mapping *slots* to *ids*, is now an unordered *set* of ids.

To use an item, the player selects the *slot* as the use key, instead
of pressing U. Since all the items in that slot are the same it uses
the first one.

There is no longer a way to drop an item. The inventory space is
unlimited. I'll have to adjust game balance later.

The UI for displaying a menu of inventory items (for Use, Drop) is no
longer needed so I removed it.